Conflicts:
	Goobi/config/goobi_config.properties
	Goobi/src/de/sub/goobi/export/dms/AutomaticDmsExportWithoutHibernate.java
	Goobi/src/de/sub/goobi/export/dms/ExportDms.java
	Goobi/src/de/sub/goobi/export/download/ExportMets.java
	Goobi/src/de/sub/goobi/forms/ProzesskopieForm.java
	Goobi/src/de/sub/goobi/forms/SpracheForm.java
	Goobi/src/de/unigoettingen/sub/search/opac/ConfigOpacCatalogueBeautifierElement.java
	Goobi/src/messages/messages_de.properties
	Goobi/src/messages/messages_en.properties
	Goobi/src/org/goobi/production/plugin/interfaces/IOpacPlugin.java

+++ Goobi/src/de/sub/goobi/export/dms/AutomaticDmsExportWithoutHibernate.java
@@ -31,6 +31,7 @@ import java.io.File;
 import java.io.IOException;
 
 import org.apache.commons.configuration.ConfigurationException;
+import org.apache.commons.io.FilenameUtils;
 import org.apache.log4j.Logger;
 
 import ugh.dl.DocStruct;
@@ -54,6 +55,7 @@ import de.sub.goobi.helper.exceptions.DAOException;
 import de.sub.goobi.helper.exceptions.ExportFileException;
 import de.sub.goobi.helper.exceptions.SwapException;
 import de.sub.goobi.helper.exceptions.UghHelperException;
+import de.sub.goobi.helper.tasks.EmptyTask;
 import de.sub.goobi.metadaten.MetadatenVerifizierungWithoutHibernate;
 import de.sub.goobi.metadaten.copier.CopierData;
 import de.sub.goobi.metadaten.copier.DataCopier;
@@ -70,6 +72,13 @@ public class AutomaticDmsExportWithoutHibernate extends ExportMetsWithoutHiberna
 	private boolean exportFulltext = true;
 	private FolderInformation fi;
 	private ProjectObject project;
+
+	/**
+	 * The field task holds an optional task instance whose progress will be
+	 * updated and whom errors will be passed to to be visible in the task
+	 * manager screen if it’s available.
+	 */
+	private EmptyTask task;
 	
 	public final static String DIRECTORY_SUFFIX = "_tif";
 
@@ -140,6 +149,9 @@ public class AutomaticDmsExportWithoutHibernate extends ExportMetsWithoutHiberna
 			gdzfile = newfile;
 
 		} catch (Exception e) {
+			if (task != null) {
+				task.setException(e);
+			}
 			Helper.setFehlerMeldung(Helper.getTranslation("exportError") + process.getTitle(), e);
 			myLogger.error("Export abgebrochen, xml-LeseFehler", e);
 			return false;
@@ -216,6 +228,9 @@ public class AutomaticDmsExportWithoutHibernate extends ExportMetsWithoutHiberna
 			if (!benutzerHome.exists()) {
 				benutzerHome.mkdir();
 			}
+			if (task != null) {
+				task.setProgress(1);
+			}
 		}
 
 		/*
@@ -229,7 +244,13 @@ public class AutomaticDmsExportWithoutHibernate extends ExportMetsWithoutHiberna
 			} else if (this.exportFulltext) {
 				fulltextDownload(process, benutzerHome, atsPpnBand, DIRECTORY_SUFFIX);
 			}
+			
+			directoryDownload(process, zielVerzeichnis);
+			
 		} catch (Exception e) {
+			if (task != null) {
+				task.setException(e);
+			}
 			Helper.setFehlerMeldung("Export canceled, Process: " + process.getTitle(), e);
 			return false;
 		}
@@ -241,6 +262,9 @@ public class AutomaticDmsExportWithoutHibernate extends ExportMetsWithoutHiberna
 		 * --------------------------------
 		 */
 		if (this.project.isUseDmsImport()) {
+			if (task != null) {
+				task.setWorkDetail(atsPpnBand + ".xml");
+			}
 			if (MetadataFormat.findFileFormatsHelperByName(this.project.getFileFormatDmsExport()) == MetadataFormat.METS) {
 				/* Wenn METS, dann per writeMetsFile schreiben... */
 				writeMetsFile(process, benutzerHome + File.separator + atsPpnBand + ".xml", gdzfile, false);
@@ -265,9 +289,10 @@ public class AutomaticDmsExportWithoutHibernate extends ExportMetsWithoutHiberna
 					Helper.deleteDir(successFile);
 				}
 			}
-			// return ;
 		}
-	
+		if (task != null) {
+			task.setProgress(100);
+		}
 		return true;
 	}
 
@@ -296,9 +321,6 @@ public class AutomaticDmsExportWithoutHibernate extends ExportMetsWithoutHiberna
 	public void fulltextDownload(ProcessObject myProzess, File benutzerHome, String atsPpnBand, final String ordnerEndung) throws IOException,
 			InterruptedException, SwapException, DAOException {
 
-		// Helper help = new Helper();
-		// File tifOrdner = new File(myProzess.getImagesTifDirectory());
-
 		// download sources
 		File sources = new File(fi.getSourceDirectory());
 		if (sources.exists() && sources.list().length > 0) {
@@ -309,9 +331,11 @@ public class AutomaticDmsExportWithoutHibernate extends ExportMetsWithoutHiberna
 			}
 			File[] dateien = sources.listFiles();
 			for (int i = 0; i < dateien.length; i++) {
-				File meinZiel = new File(destination + File.separator
-						+ dateien[i].getName());
-				Helper.copyFile(dateien[i], meinZiel);
+				if(dateien[i].isFile()) {
+					File meinZiel = new File(destination + File.separator
+							+ dateien[i].getName());
+					Helper.copyFile(dateien[i], meinZiel);
+				}
 			}
 		}
 		
@@ -327,84 +351,20 @@ public class AutomaticDmsExportWithoutHibernate extends ExportMetsWithoutHiberna
 					}
 					File[] files = dir.listFiles();
 					for (int i = 0; i < files.length; i++) {
-						File target = new File(destination + File.separator + files[i].getName());
-						Helper.copyFile(files[i], target);
+						if(files[i].isFile()) {
+							File target = new File(destination + File.separator + files[i].getName());
+							Helper.copyFile(files[i], target);
+						}
 					}
 				}
 			}
 		}
-		
-		
-//		File sources = new File(fi.getSourceDirectory());
-//		if (sources.exists()) {
-//			File destination = new File(benutzerHome + File.separator
-//					+ atsPpnBand + "_src");
-//			if (!destination.exists()) {
-//				destination.mkdir();
-//			}
-//			File[] dateien = sources.listFiles();
-//			for (int i = 0; i < dateien.length; i++) {
-//				File meinZiel = new File(destination + File.separator
-//						+ dateien[i].getName());
-//				Helper.copyFile(dateien[i], meinZiel);
-//			}
-//		}
-//		
-//		
-//		File txtFolder = new File(this.fi.getTxtDirectory());
-//		if (txtFolder.exists()) {
-//			File destination = new File(benutzerHome + File.separator + atsPpnBand + "_txt");
-//			if (!destination.exists()) {
-//				destination.mkdir();
-//			}
-//			File[] dateien = txtFolder.listFiles();
-//			for (int i = 0; i < dateien.length; i++) {
-//				File meinZiel = new File(destination + File.separator + dateien[i].getName());
-//				Helper.copyFile(dateien[i], meinZiel);
-//			}
-//		}
-//
-//		File wordFolder = new File(this.fi.getWordDirectory());
-//		if (wordFolder.exists()) {
-//			File destination = new File(benutzerHome + File.separator + atsPpnBand + "_wc");
-//			if (!destination.exists()) {
-//				destination.mkdir();
-//			}
-//			File[] dateien = wordFolder.listFiles();
-//			for (int i = 0; i < dateien.length; i++) {
-//				File meinZiel = new File(destination + File.separator + dateien[i].getName());
-//				Helper.copyFile(dateien[i], meinZiel);
-//			}
-//		}
-//
-//		File pdfFolder = new File(this.fi.getPdfDirectory());
-//		if (pdfFolder.exists()) {
-//			File destination = new File(benutzerHome + File.separator + atsPpnBand + "_pdf");
-//			if (!destination.exists()) {
-//				destination.mkdir();
-//			}
-//			File[] dateien = pdfFolder.listFiles();
-//			for (int i = 0; i < dateien.length; i++) {
-//				File meinZiel = new File(destination + File.separator + dateien[i].getName());
-//				Helper.copyFile(dateien[i], meinZiel);
-//			}
-//		}
 	}
 
 	public void imageDownload(ProcessObject myProzess, File benutzerHome, String atsPpnBand, final String ordnerEndung) throws IOException,
 			InterruptedException, SwapException, DAOException {
 		/*
-		 * -------------------------------- erstmal alle Filter
-		 * --------------------------------
-		 */
-		// FilenameFilter filterTifDateien = new FilenameFilter() {
-		// public boolean accept(File dir, String name) {
-		// return name.endsWith(".tif");
-		// }
-		// };
-
-		/*
-		 * -------------------------------- dann den Ausgangspfad ermitteln
+		 * -------------------------------- den Ausgangspfad ermitteln
 		 * --------------------------------
 		 */
 		File tifOrdner = new File(this.fi.getImagesTifDirectory(true));
@@ -428,7 +388,11 @@ public class AutomaticDmsExportWithoutHibernate extends ExportMetsWithoutHiberna
 				 */
 				Benutzer myBenutzer = (Benutzer) Helper.getManagedBeanValue("#{LoginForm.myBenutzer}");
 				try {
-                    FilesystemHelper.createDirectoryForUser(zielTif.getAbsolutePath(), myBenutzer.getLogin());				} catch (Exception e) {
+					FilesystemHelper.createDirectoryForUser(zielTif.getAbsolutePath(), myBenutzer.getLogin());
+				} catch (Exception e) {
+					if (task != null) {
+						task.setException(e);
+					}
 					Helper.setFehlerMeldung("Export canceled, error", "could not create destination directory");
 					myLogger.error("could not create destination directory", e);
 				}
@@ -438,10 +402,58 @@ public class AutomaticDmsExportWithoutHibernate extends ExportMetsWithoutHiberna
 
 			File[] dateien = tifOrdner.listFiles(Helper.dataFilter);
 			for (int i = 0; i < dateien.length; i++) {
-				File meinZiel = new File(zielTif + File.separator + dateien[i].getName());
-				Helper.copyFile(dateien[i], meinZiel);
+				if (task != null) {
+					task.setWorkDetail(dateien[i].getName());
+				}
+				if(dateien[i].isFile()) {
+					File meinZiel = new File(zielTif + File.separator + dateien[i].getName());
+					Helper.copyFile(dateien[i], meinZiel);
+				}
+				if (task != null) {
+					task.setProgress((int) ((i + 1) * 98d / dateien.length + 1));
+					if (task.isInterrupted()) {
+						throw new InterruptedException();
+					}
+				}
+			}
+			if (task != null) {
+				task.setWorkDetail(null);
 			}
 		}
+	}
 
+	/**
+	 * starts copying all directories configured in goobi_config.properties parameter "processDirs" to export folder 
+	 * 
+	 * @param myProzess the process object
+	 * @param zielVerzeichnis the destination directory
+	 * @throws IOException
+	 */		
+	private void directoryDownload(ProcessObject myProzess, String zielVerzeichnis) throws IOException{
+	
+		String[] processDirs = ConfigMain.getStringArrayParameter("processDirs");
+		
+		for(String processDir : processDirs) {
+		
+			File srcDir = new File(FilenameUtils.concat(fi.getProcessDataDirectory(), processDir.replace("(processtitle)", myProzess.getTitle())));
+			File dstDir = new File(FilenameUtils.concat(zielVerzeichnis, processDir.replace("(processtitle)", myProzess.getTitle())));
+		
+			if(srcDir.isDirectory()) {
+				Helper.copyDir(srcDir, dstDir);
+			}
+		}
+	}
+
+	/**
+	 * The method setTask() can be used to pass in a task instance. If that is
+	 * passed in, the progress in it will be updated during processing and
+	 * occurring errors will be passed to it to be visible in the task manager
+	 * screen.
+	 * 
+	 * @param task
+	 *            task object to submit progress updates and errors to
+	 */
+	public void setTask(EmptyTask task) {
+		this.task = task;
 	}
 }
